+++ src/main/java/ai/mcpdirect/studio/tool/openapi/OpenAPIToolProvider.java
@@ -3,6 +3,7 @@ package ai.mcpdirect.studio.tool.openapi;
 import ai.mcpdirect.studio.dao.entity.OpenAPIServer;
 import ai.mcpdirect.studio.tool.AITool;
 import ai.mcpdirect.studio.tool.util.AIToolProvider;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.Collection;
 import java.util.List;
@@ -10,6 +11,12 @@ import java.util.Map;
 
 public class OpenAPIToolProvider extends OpenAPIServer implements AIToolProvider {
     public OpenAPIToolProvider(OpenAPIServerConfig config){
+
+    }
+
+    public void refreshTools() {
+    }
+    public void close(){
 
     }
     @Override
@@ -26,4 +33,5 @@ public class OpenAPIToolProvider extends OpenAPIServer implements AIToolProvider
     public String callTool(String name, Map<String, Object> parameters) {
         return "";
     }
+
 }